backend/requirements.txt
# ===== FastAPI and Server =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ===== Database (PostgreSQL + PostGIS + pgvector) =====
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.14.3
pgvector==0.2.5

# ===== Authentication & Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# ===== SOTA Computer Vision (EfficientNetV2 + CLIP) =====
torch==2.9.1
torchvision==0.24.1
timm==0.9.12
ftfy==6.1.3
regex==2023.12.25
git+https://github.com/openai/CLIP.git

# ===== SOTA NLP (GLiNER for Zero-Shot NER) =====
gliner==0.2.8

# ===== Image Processing =====
Pillow==11.0.0
opencv-python==4.10.0.84

# ===== Geospatial =====
shapely==2.0.6

# ===== Utilities =====
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1
numpy==2.2.1

# ===== Testing =====
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
